+++ app/auth/admin/group_views.py
@@ -5,9 +5,10 @@ from flask_babel import lazy_gettext,gettext
 from flask_login import login_required
 
 from app.utils import admin_required, crypt
-from app.data.models.group import Group
+from app.data.models import Group, User
 from app.public.forms import EditGroupForm
 from . import admin
+import json
 
 
 @admin.route('/group/list', methods=['GET', 'POST'])
@@ -58,3 +59,19 @@ def group_delete(str_hash):
     group.delete()
     flash(gettext('Group {nazev} deleted').format(nazev=group.nazev),'success')
     return redirect(url_for('.group_list'))
+
+
+class CustomEncoder(json.JSONEncoder):
+    def default(self, obj):
+        if isinstance(obj, User):
+            return obj.to_json()
+        return json.JSONEncoder.default(self, obj)
+
+
+@admin.route('/group/add/<str_hash>', methods=['GET', 'POST'])
+def group_add_user(str_hash):
+    id = int(float(crypt(str_hash, decrypt=True)))
+    group = Group.query.filter_by(id=id).first_or_404()
+    users = User.query.all()
+    pole = json.dumps(users, cls=CustomEncoder)
+    return render_template('group_add_users.html', pole=pole)
