+++ app/auth/admin/templates/group_add_users.html
@@ -29,16 +29,16 @@
 
 {% block body %}
     <div>
-        <div style="width: 45%; float: left;">
+        <div style="width: 45%; display: inline-block;">
             <table class="table table-bordered" id="panel1">
 
             </table>
         </div>
-        <div style="width: 10%; float: left; text-align: center">
+        <div style="width: 9%; display: inline-block; text-align: center;">
             <button id="btn-right" class="btn btn-default"><span class="glyphicon glyphicon-arrow-right"></span></button><br>
             <button id="btn-left" class="btn btn-default"><span class="glyphicon glyphicon-arrow-left"></span></button>
         </div>
-        <div style="width: 45%; float: left;">
+        <div style="width: 45%; display: inline-block;">
             <table class="table" id="panel2">
                 <tr><td>(empty)</td></tr>
             </table>
@@ -51,10 +51,16 @@
         var users = {{ pole|safe }}
 
         $(document).ready(function() {
-            var table = $('#panel1').DataTable( {
+            var table1 = $('#panel1').DataTable( {
                 data: users,
                 columns: [
-                    { title: "Name" }
+                    { title: "{% trans %}Name{% endtrans %}" }
+                ]
+            } );
+            var table2 = $('#panel2').DataTable( {
+                data: [],{# users, #}
+                columns: [
+                    { title: "{% trans %}Name{% endtrans %}" }
                 ]
             } );
 
@@ -62,12 +68,16 @@
                 $(this).toggleClass('selected');
             } );
 
+            $('#panel2 tbody').on( 'click', 'tr', function () {
+                $(this).toggleClass('selected');
+            } );
+
             $('#btn-right').click( function () {
-                console.log(table.rows('.selected').data());
+                console.log(table1.rows('.selected').data());
             } );
 
             $('#btn-left').click( function () {
-                console.log(table.rows('.selected').data());
+                console.log(table2.rows('.selected').data());
             } );
 
         } );
